% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_econometric_model.R
\name{choose_best_econometric_model}
\alias{choose_best_econometric_model}
\title{With the usage of the specific dataset and given target variable by the user
regarding the explained variable, a function automatically selects the best
form of the model based on significance values and forward selection}
\usage{
choose_best_econometric_model(modeling_dataframe, general_formula, target_var)
}
\arguments{
\item{modeling_dataframe}{table chosen for modeling}

\item{general_formula}{a general modeling formula}

\item{target_var}{target variable to model}
}
\value{

}
\description{
With the usage of the specific dataset and given target variable by the user
regarding the explained variable, a function automatically selects the best
form of the model based on significance values and forward selection
}
