% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_month_df.R
\name{create_month_df}
\alias{create_month_df}
\title{Create aggregated month dataframe based on daily features}
\usage{
create_month_df(path_to_month_df = "")
}
\arguments{
\item{path_to_day_df}{either path relative to server.R file
or complete path to day dataframe}

\item{write_month_df_to_file}{logical whether to save final dataframe
to Input_files_01 folder}
}
\value{

}
\description{
Create aggregated month dataframe based on daily features
}
