% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_day.R
\name{eda_day}
\alias{eda_day}
\title{Create complex exploratory data analysis report for
daily rentals.}
\usage{
eda_day()
}
\arguments{
\item{path_to_day_df}{path to dataframe with daily statistics}
}
\value{

}
\description{
Create complex exploratory data analysis report for
daily rentals.
}
