% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_prety_name.R
\name{get_target_pretty_name}
\alias{get_target_pretty_name}
\title{Returns variable pretty name}
\usage{
get_target_pretty_name(target_var)
}
\arguments{
\item{target_var}{target variable, one of casual, registered or total}
}
\value{

}
\description{
Returns variable pretty name
}
