% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barchart.R
\name{plot_barchart}
\alias{plot_barchart}
\title{Function yield a barchart fo chosen rental variable}
\usage{
plot_barchart(group_var, target_var)
}
\arguments{
\item{group_var}{either yr or holiday, produces barplot for this group
where values represent rental counts per day}

\item{target_var}{variable to display on y axis in barcharts}
}
\value{

}
\description{
Function yield a barchart fo chosen rental variable
}
