% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_rentals.R
\name{predict_rentals}
\alias{predict_rentals}
\title{Complete function to load a forecasting model and yield the plot of
forecast vs actual values. We use bike rentals data for two years. We
have built the forecast prediction for the next quarter as well as
for the next 30 days}
\usage{
predict_rentals(granularity_level, target_var)
}
\arguments{
\item{granularity_level}{either monthly or daily}
}
\value{
pretty plot of bike rentals prediction
}
\description{
Complete function to load a forecasting model and yield the plot of
forecast vs actual values. We use bike rentals data for two years. We
have built the forecast prediction for the next quarter as well as
for the next 30 days
}
