% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_day_df.R
\name{preprocess_day_df}
\alias{preprocess_day_df}
\title{Reads the day dataframe and performs all necessary steps in  particular
splitting dataframe into separate dataframes
(datetime, weather, normalized regressors and outcome)}
\usage{
preprocess_day_df()
}
\arguments{
\item{path_to_day_df}{either path relative to server.R file
or complete path to day dataframe}
}
\value{
list
}
\description{
Reads the day dataframe and performs all necessary steps in  particular
splitting dataframe into separate dataframes
(datetime, weather, normalized regressors and outcome)
}
